Test a worker config locally with container
Test within the gitlab CI in the unpriviledged containers where we can
not change the hostname from the outside by just adjusting the salt
recipes to include the current host as a worker however that is a
generic one. We can extend that to cover other special workers as well,
e.g. in parallel test jobs.