Add osd-admins to all current alerts
Until now we relied on the "default notification channel" configured
inside grafana. Since one cannot disable the default channel on panels
where it is not relevant (e.g. panels with automated action triggers),
this causes a lot of overhead on the mailing list. Therefore we disable
it now as default but add it to every configured alert explicitly.