Prevent generic logrotate service failing by additional openqa logrotate service+timer
Overriding the generic logrotate service causes the systemd service to
fail except on midnight because it correctly skips the log rotation for
the services which have a daily profile however the exit code of
logrotate still reflects a failure in this case. Adding an explicit
additional logrotate-openqa service and timer and keeping the generic
logrotate service for the daily execution prevents this.