Extend NVMe preparation to be applicable in general
Tested locally in a container equivalent to CI tests and also on
openqaworker13 which has one NVMe. The salt states describe the desired
end state but can not be applied completely automatically, i.e. errors
will occur that need to be fixed also by reboot of the machines.

For now the state can be applied for all machines with the role
"nvme_store" so that we can apply the state explicitly where
prerequisites are fulfilled. After we migrated enough machines we can
apply the state for all workers.

Related progress issue: https://progress.opensuse.org/issues/19238