Fix "Currently failing services" table in grafana
Telegraf writes the number of failed services on each host into our
InfluxDB. Our current query takes the whole dataset and displays each
machine which last reported at least one failed service. If a machine
goes down and the last failed service count was >1, this machine will
always show up in the table. This commit takes the time filter into
account and sets a time filter override of always 15 minutes to the
table in grafana resulting in always having machines with at least 1
failed service but not older then 15 minutes in the list.