Improve deployment if minions are down
Previously we fail the whole deployment pipeline if one single minion is
down (this is due to the return value of the salt calls). With
--hide-timeout, salt ignores these minions and exits cleanly. In case of
other issues salt would still exit with != 0.
Additionaly, at the beginning of each deployment, the pipeline now
prints which minions are timing out and displays it to the operator.