Add installation of the grafana image renderer
This adds an automatic installation of the grafana image renderer plugin
onto our monitoring host. It uses the administration tool from grafana
to install https://grafana.com/grafana/plugins/grafana-image-renderer if
it is not already installed on the target host. To activate the plugin,
the grafana-service gets restarted after installation.