Fix ssh based rules failing for environments with no ssh
Commonly openssh is installed in many environments, e.g. all our remote
systems as well as test containers. For the sake of completeness the
dependency on the openssh package should be specified though. This also
fixes a problem when testing manually in an environment where the
package is not yet installed, e.g. bare containers with less packages
installed.