nvme_store: Add support for machines with only a single NVMe
Currently openqaworker8 and openqaworker9 have just a single NVMe, no
other devices, e.g. no disk or SSD. Previously we skipped the
initialization and let this for manual control. With this commit we can
also dynamically handle single NVMe machines.

I did the distinction within the systemd service and not based on salt
grains to prevent a mismatch in case of hardware changes, e.g when one
powers down the machine and changes the setup of NVMes or SSDs without
giving salt a chance to update the grains cache. Instead the service
decides at runtime which devices/partitions can be used for the RAID0
initialization.

Tested on openqaworker8.

Related progress issue: https://progress.opensuse.org/issues/19238