Ignore incomplete jobs with reason 'quit'
In this case the worker has been restarted within a reasonable
time (systemd had not had to kill it). This usually happens during
deployment or when the worker is restarted/stopped for other
reasons. We can ignore these incompletes. If the duplication fails
the reason is turned into 'api failure: ...' so we will not miss
jobs which could not be duplicated.

See https://github.com/os-autoinst/scripts/pull/18