grafana: Save some wait time for automatic actions
For the three current aarch64 machines within osd we have automatic
actions power cycling the machines triggered by alerts after 1h of
downtime. We previously selected 1h to provide a window for manual
debugging however we learned that we can not interact with the machine
at all during this time and we are better off monitoring the traffic on
the serial port automatically.

Related progress issue: https://progress.opensuse.org/issues/41882