monitoring: Save 'status overview' dashboard including new 'availability' panel
The "status overview" dashboard was so far not managed by salt. This
commit adds the complete dashboard to be provisioned over salt as I also
added another panel that calculates the "availability" of the openQA
webUI on openqa.suse.de. This is motivated by the plan within SUSE to
provide more metrics in general.

Related progress issue: https://progress.opensuse.org/issues/68233