Ensure grains are synced as well as refreshed
After adding a custom grain that is synced on every high state but to
ensure it is also available for an initial high state it seems to be the
safer approach to ensure it is synced in before. I tried an approach as
described on
https://docs.saltstack.com/en/latest/topics/reactor/index.html#minion-start-reactor
but that did not seem to have any effect. Probably because I forgot to
restart the salt master service ;)