Add complete monitoring dashboard for "age of scheduled jobs"
This dashboard includes both abstracted singlestat panels for the
generic status as well as graphs with two alerts for more low-level
details. We might want to remove some of the low-level panels if they
are not helpful as well as move the abstracted singlestat panels to the
top-level "status" dashboard.

Related progress issue: https://progress.opensuse.org/issues/65975