monitoring: Prevent "only running but no scheduled" causing false alert
Previously we looked at "not done and not cancelled" hence taking
running, assigned, scheduled, blocked into account in the same way.
However the graphs and alerts mainly took the case in consideration that
there are scheduled jobs regardless of any running. This triggered false
alerts in case when the last scheduled job in a long list was picked up
and turned to running.

This commit changes the query to now only look at scheduled jobs (which
includes the blocked ones) and also "assigned" as they have a very
similar meaning, i.e. jobs that have not been actually picked up by
workers.