Remove openvswitch workaround obsoleted by os-autoinst update
https://github.com/os-autoinst/os-autoinst/pull/1454 should ensure that
os-autoinst-openvswitch.service can be started even earlier by waiting
internally for the right bridge device as we still had problems with the
network stack to be not fully initialized. The second problem that the
above PR fixes is that the bridge device can be configured so "br1"
might not even be right device in all cases.

Related progress issue: https://progress.opensuse.org/issues/59300