openqa-trigger-from-ibs: Delete duplicate pkg dependencies
f3857ce introduced a partial copy-paste of the package dependencies for
openqa-trigger-from-ibs-plugin which does not help as the former package
needs to be installed first anyway. The dependency for SUSE internal SSL
certificates to access the gitlab repo was missing as well so this is
anyway incomplete. Also, the additional pkg.installed instruction yields
another time consuming zypper call.