CI: Move deploy definitions to a separate file for easier inclusion
Merge requests against salt-pillars-openqa try to execute tests from
.gitlab-ci.yml from this repo. We already exclude the origin repository
because the tests should not be called, only the deploy step. So we can
move the deploy definitions into its own file so that we can include
that file in both the local .gitlab-ci.yml file as well as in
openqa-pillars-openqa.

Related progress issue: https://progress.opensuse.org/issues/70918