monitoring: Fix potential wrong lookup of ping value with 'last' instead of 'distinct'
I observed that sometimes a machine would be reported as OFFLINE even
though it was online at this time. Hopefully the use of 'last' can fix
this. This might also be more efficient by looking up just the last
value, not needing to iterate over multiple values in the database.