monitoring: Bump up limits for CPU usage "IOwait too high"
Relaxing limits further by switching alert level from 42 to 50 for CPU
usage and extending the time range from 30s to 5m.

Also adding a description for the alert with a suggestion what to check
as well as a ticket reference for a likely explanation what changed in
the past.

Related progress issue: https://progress.opensuse.org/issues/69664