monitoring: remove idle alert
While it can be a critical sign if a worker is fully utilized the alert
as it is yields almost no benefit as you cannot really act on it.
Besides that we don't really care if a system is fully utilized, we
rather want this. The real problem starts if a specific job queue is
getting too long and we have this covered in a separate alert rule.