Fix racy bind mounts with proper systemd dependencies
There was a problem that after reboot the bind mount point was created
before /assets was mounted.
systemd does not look at the ordering information in /etc/fstab which
was not reliable for us anyway so we need to specify proper
dependencies.

Related progress issue: https://progress.opensuse.org/issues/73501