Add better visibility for missing metrics in worker graphs
Previously we connected points if data was missing. AFAICT this was done
to avoid "dotted graphs" but also resulted in misleading graphs if data
was missing for other reasons (e.g. if the machine is actually down). To
avoid the dotted graphs, the minimum interval got adjusted to the
scraping interval (see: openqa/telegraf-worker.conf#L3) which in turn
allows us to set the null values to actual null values. This then renders
a proper graph with gaps if data is missing.