Add custom grain for default network interface
This is based on recommendations from
https://lemarchand.io/saltstack-and-internal-network-interfaces/
and allows us to determine which network interface is the default one
used for IPv4. This information can later be used for various other
operations. Currently this is a very basic implementation which might
break if you have multiple interfaces with IPs in the subnet
10.160.0.0/13 which should not be the case inside SUSE.
Example output of this grain:

```
salt '*' grains.get "default_interface"
1b137.qa.suse.de:
    br0
```