Ensure proper static and transient hostnames for all workers
As a fallback if the initial network configuration takes a bit we should
have the static hostnames properly configured. This likely also makes
logs a bit more readable when the hostname in the log is something we
"know".

Following
https://docs.saltstack.com/en/latest/ref/states/all/salt.states.network.html
it should be very much possible to set the hostname with something like

```
system:
  network.system:
    - hostname: {{ grains['host'] }}
    - apply_hostname: True
    - retain_settings: True
```

but this showed an error:

```
          ID: system
    Function: network.system
      Result: False
     Comment: 'ip.get_network_settings'
     Started: 16:30:30.656685
    Duration: 0.928 ms
     Changes:
```

so I am just doing:

```
/etc/hostname:
  file.managed:
    - contents: {{ grains['host'] }}
```

instead which seems to not apply the change *immediately* but some
seconds after `hostnamectl` reports the right hostname.

An actual alternative could be to just have an empty file /etc/hostname
which some workers had in before as well.

Related progress issue: https://progress.opensuse.org/issues/76786