Delete non-actionable alert about "incompletes"
We have an alert about "not restarted incompletes" that is "actionable"
and if it is really for just restarting these jobs. But plain
"incompletes of last 24h" would only vanish from alerting again if we
wait long enough so that older incompletes go out of the 24h window or
if we delete jobs which we normally do not want to do to prevent loosing
information too early.

This commit deletes the alert. The panel to visualize the data is
still preserved.

Related progress issue: https://progress.opensuse.org/issues/80118