Ensure telegraf service to be enabled to start immediately on boot
I realized that on both openqaworker6 and openqaworker10 telegraf was
not running immediately after boot. The systemd service were not
enabled.

I think this can explain why the weekly reboot caused a delay of longer
than 1h until hosts came back because in salt we never ensure that the
service is enabled, we just ensure "running", not sure when salt is even
triggered and I wonder why they even show up after 1h.

This commit changes telegraf service to not only be ensured to be
"running" in runtime when salt is trigger a high state but also ensuring
a persistent configuration to automatically start up the services on
boot.

Related progress issue: https://progress.opensuse.org/issues/87883