Raise system level dbus connection limit
As we have a lot of services running on our workers it can happen that
the default connection limit is reached and services fail to start (e.g.
see https://progress.opensuse.org/issues/88225 for an example). Our
servers have enough power to handle more of those connections and
therefore raising the total limit seems to be a reasonable fix. If this
limit set here needs to be bumped again in the near future, we need to
investigate if we have danling connections somewhere.