Weaken NTP offset alert; also consider negative offsets
With this change the alert only fires if the offset stays outside the range
of -6 seconds to 6 seconds for more than 15 minutes. Before the range
was -∞ to 1 second and the average over a time of 5 seconds was considered.

We've seen offsets within the new range in production from time to time and
I am not aware that the offset caused any real problems although the offset
was present for multiple hours.

See https://progress.opensuse.org/issues/88450