Remove `openqa-worker.target: service.running` to prevent activating unwanted service
Despite `enable: False` this apparently¹ activates `openqa-worker.target`
which would start the regular worker service (and we want
`openqa-worker-auto-restart@.service`).

Even if we wanted to use the regular worker service this would lead to
activating unwanted worker slots because `openqa-worker.target` does not
care about the number of slots configured in salt and simply starts worker
slots for the present pool directories (at the time of the last systemd
generator run).

Related tickets:
* https://progress.opensuse.org/issues/63874
* https://progress.opensuse.org/issues/80908

---

¹ https://docs.saltproject.io/en/latest/ref/states/all/salt.states.service.html#salt.states.service.running