Move NVMe creation into a separate script
This should help to proceed with https://progress.opensuse.org/issues/88191
and https://progress.opensuse.org/issues/80482. The code/behavior is not
supposed to change by this refactoring commit.

Note that `RequiresMountsFor=/usr/local/bin` is required because this
directory might not be on the root mount point. (In fact, in the production
workers `/usr/local` seems to be on a separately mounted btrfs subvolume.)