Fix openSUSE Leap 15.2 PowerPC workers stuck in petitboot
Within petitboot (or skiboot) pb-discover parses the grub config from
the installed Linux system to find the correct boot entries. And as the
logfile says pb-discover fails to parse the grub config in line 237
which in our system corresponds to line 241 as reported in
https://bugzilla.opensuse.org/show_bug.cgi?id=1174166#c1 . The bug has
been set to RESOLVED FIXED for Factory and SLE15-SP3 (and hence openSUSE
Leap 15.3) only, not openSUSE Leap 15.2 as we run it.

Added a comment to the bug report with
https://bugzilla.opensuse.org/show_bug.cgi?id=1174166#c14
and created a maintenance request
https://build.opensuse.org/request/show/882581

With this commit we apply a workaround from Linux hot-patching our grub
as well.

Related progress issue: https://progress.opensuse.org/issues/68053