Prevent premature NTP alerts when workers take longer to reboot
Multiple alerts for qa-power8-5-kvm as visible on
https://stats.openqa-monitor.qa.suse.de/d/WDQA-Power8-5-kvm/worker-dashboard-qa-power8-5-kvm?editPanel=65102&tab=alert&viewPanel=65102&orgId=1&from=1617299599160&to=1617310314965
where the alerts seem to coincide with reboots that I triggered as an
experiment. It seems that the NTP offset is exceeding a threshold as
there are many values missing due to reboot. Likely we can live with a
longer pending time to prevent this problem.

This commit bumps the pending time for the NTP worker alerts to 20
minutes which should provide stable NTP values after a longer
multi-minute reboot time which is likely to be expected especially from
our PowerPC machines.

Related progress issue: https://progress.opensuse.org/issues/90635