Ensure salt-minion is enabled to cover reboot-handling
As we trigger automatic reboots of salt-managed machines we should
ensure that any encountered salt-minion is also enabled to start after
reboot.

This fixes a problem as observed with storage.qa.suse.de that was added
to the salt-managed infrastructure but the salt-minion was never enabled
to be running after reboot.

Related progress issue: https://progress.opensuse.org/issues/90746