Fix `_grains/default_interface.py` for Python 3
This should fix one of the backtraces we see in production (see
https://progress.opensuse.org/issues/94994).

The problem has also been reported upstream (see
https://github.com/saltstack/salt/issues/58860) and apparently the solution
is to use `salt.utils.network` directly instead of `salt.modules.network`.
This change implements that.

---

Tested by appending `print(default_interface())` at the end of the grain
and invoking it locally via `python3 _grains/default_interface.py`.

Before the change an error message similar to the one from production
occurs:

```
Traceback (most recent call last):
  File "_grains/default_interface.py", line 24, in <module>
    print(default_interface())
  File "_grains/default_interface.py", line 10, in default_interface
    internalip = salt.modules.network.ip_addrs(cidr='10.160.0.0/13')[0]
  File "/usr/lib/python3.8/site-packages/salt/modules/network.py", line 1278, in ip_addrs
    addrs = __utils__["network.ip_addrs"](
NameError: name '__utils__' is not defined
```

With the change it returns the interface used to connect to SUSE's internal
network (which seems to be the intention of the code):

```
{'default_interface': 'tun0'}
```

(Since I'm only connected via VPN to the internal network, `tun0` is the
correct interface in my local test.)