Tweak asset cleanup value to allow to use a bit more available space
https://monitor.qa.suse.de/d/WebuiDb/webui-summary?viewPanel=74&orgId=1&from=now-1y&to=now
shows that with the space-aware cleanup since multiple months we are savely
below the alert threshold of 90% and can tweak to go a little bit further to
the red line without touching it. A little peak during that period was touching
84% so I guess we can move our limit from 80 to actually 84 to use some more of
the available space for assets.

Related progress issue: https://progress.opensuse.org/issues/70885