Consistently configure openQA logging to file
We had the logging to file configured but then commit 8686c95 from 2017
removed it with the commit message "Do not require file logging" outside
any merge request so it is hard to find out what the actual intention
was. This commit re-enables logging to file again in the hope to give
other system related logs a chance to end up in the system journal.

Related progress issue: https://progress.opensuse.org/issues/96551