Fix Minion workers alert
The last attempt still wasn't good enough because the alert could easily
fire if the worker was switching between active and inactive for too long.
I'm now using a third query `max(active) + max(inactive)` for the alert
which would fire if it is outside the range of 1 to 2. That should be more
reliable.

See https://progress.opensuse.org/issues/96380