monitoring: Also exclude 'tests died' incompletes same as in os-autoinst/scripts
In
https://github.com/os-autoinst/scripts/blob/master/openqa-monitor-incompletes#L6
we already opted to exclude also "tests died" and "abandoned". So we
could do the same here. I do this change mainly because of instances
where we found "tests died" because of incompatible test distribution
specific test schedule changes where we should not need to act upon. For
"abandoned" we have automatic cloning within openQA so it should not
make a difference as long as auto-cloning works. As soon as it fails we
should still be alerted about that.

https://progress.opensuse.org/issues/99153