Relax openvswitch init timeout due to severe slowdown in the past days
During the past days in particular on openqaworker-arm-3 which crashes
and reboots often we see recurring failures of the service
os-autoinst-openvswitch due to the network not being fully up and
available even after a waiting time of 5 minutes. This seems to be a
severe configuration problem or wicked bug. However as observable in the
system journal eventually the network does show up so waiting for even
longer should help in such cases. Hence this commit bumps the timeout
for initialization of os-autoinst-openvswitch from 5m to 15m.

Related progress issue: https://progress.opensuse.org/issues/106594