openqa: Fix broken postgreSQL config pattern replacement
Manually testing the changes on OSD in a separate file with

```
salt-call --local -l error --state-output=changes state.sls_id postgresql-work_mem2 openqa.postgresql test=True
```

which takes about 10s each run – why is salt so slow – but ok.

Trying out the pattern manually with grep or sed was fine but with salt
the replacement ends up as "tMB". Likely the problem is that salt tries
to find a pattern group "164" instead of "1" for "\164". grep and sed
behave different. This can easily be fixed with moving the space from
the pattern to the replacement.

I tried to verify the broken state in psql with `select
pg_reload_conf();` which still shows "t" despite the config being
invalid. This is in line with
https://stackoverflow.com/questions/16333319/how-to-syntax-check-postgresql-config-files#comment101836284_55758799
which also states that this does not work. Also
https://stackoverflow.com/a/16340569 states that there is no separate
check config command. One needs to try to restart the server which
refuses to load an invalid config.

The pattern replacement is within a block `{%- if not
grains.get('noservices', False) %}` so we will not see those rules
effective in CI jobs.

Related progress issue: https://progress.opensuse.org/issues/112916