monitoring: Be more flexible with sporadically higher number of minion workers
Previous attempts for example in
https://progress.opensuse.org/issues/96380
corresponding to commits fde8933 and 3fbe3ea tried to get the accounting
right for the sum of all existing minion workers. The latter change also
changed the alert to only accept numbers within the range [1;2] which in
the past days cause alerts when the calculated number was reported to be
higher than 2. I think we can ignore any number higher than 1 and again
only alert when the sum of the max values is below 1.

Related progress issue: https://progress.opensuse.org/issues/112898