Re-try NVMe creation if the RAID array is already in use
We already try to stop an in-use RAID before. However, on `openqaworker14`
we just ran into the error that the RAID was nevertheless already in-use.
I suppose it came up just after checking to stop it so the code for
stopping it was not executed (`Stopping current RAID "/dev/md/openqa"` was
not logged but `[[ -e /dev/md/openqa ]]` was true when I manually checked
later and `mdadm --stop /dev/md/openqa` worked when I ran the command again
by invoking the script manually). So I hope this change will help.

See https://progress.opensuse.org/issues/114526#note-8