Ensure kdump is enabled on all machines, not only workers
With
https://gitlab.suse.de/openqa/salt-states-openqa/-/merge_requests/733 we
introduced automatic reboots on all machines if the kernel panics. This
could hide away problems. Therefore we should make sure to enable kdump
on all machines to get reports if the kernel panics and fix the
underlying problem. Rebooting should just be a workaround not the
solution.

Related progress issue: https://progress.opensuse.org/issues/116848