Simplify and potentially optimize openQA worker installations
Originally it was necessary to have two separate package lists as we
explicitly selected one list to be installed from the devel:openQA repo.
Since we use proper repo definitions with according priorities now we
can merge the package list. This is easier to maintain here and might
improve the performance as individual calls to package installations are
reduced.