Try to fix missing change-triggered telegraf restarts on all hosts
Originally in
https://gitlab.suse.de/openqa/salt-states-openqa/-/merge_requests/467
the change was introduced to only watch telegraf config files on webui
and worker roles but we run telegraf on all the roles hence the telegraf
service would not restart on changes on such hosts, e.g. a monitor host.
This change removes the restriction to fix that in the hope that the
originally mentioned CI problems do not reappear.