Rely on FQDNs of openQA worker to use more reliable DNS entries and IPv6
There are variables called WORKER_HOSTNAME and SUT_IP and we often use
IPv4 addresses even though we already use domain names in multiple
cases.

https://stackoverflow.com/a/54750233 recommends to avoid FQDN and in
nearly all cases our salt ID is already equivalent to the FQDN so this
can be an alternative choice but means more restrictions how we can
select the salt minion_id. I think we should just rely that each host
knows each FQDN and that DNS entries are properly working. Then we can
be independant of actual IPv4 addresses and also make use of more
efficient IPv6.

The prerequisite on the side of worker hosts is that on each host
`nslookup $(hostname)` works fine, needing proper DNS entries as well as
the according search entry for the current hostname in /etc/resolv.conf
which normally should be supplied and is already the case of all openQA
workers in the openqa.suse.de domain.

Related progress issue: https://progress.opensuse.org/issues/109241