Use our own devel repository for Grafana
After switching to unified alerts we should stay closer to upstream
because the new alerting system is still work in progress and new
releases are still expected to bring important features and bug fixes.

The maintainers of the Grafana package under `server:monitoring` are
unfortunately very slow to respond so it can take weeks to get a change
in. So at least for the time being it makes sense to use our own
repository (still trying to submit changes to `server:monitoring` of
course).