Grafana: Update apache response time panel
This migrates the "Apache Response Time" panel to the new Time Series
panel. It also changes colors of each metric (min, avg, max) to green,
yellow and red to make their difference more clear. In addition an
unused, additional query (called "A") was deleted which caused the panel
to not render longer time periods (e.g. observed with a 7 day range).