Fix customization of alert threshold for partition usage alert
This is a follow-up of 5a367a101d7ee30c35338aad1b98bd9ee534b3fd which
did not work because we must not use `grains` here. This template is
deployed on the monitoring host so `grains['host']` would resolve for
the monitoring host although we need the generichost/worker here.

Related ticket: https://progress.opensuse.org/issues/127274