monitoring: alerting: Add webUI Disk I/O time alert
I created the alert in the grafana webUI, then exported all alerts to a
YAML document and included only the section about the new alert in the
file dashboard-WebuiDb and adjusted the folder and key ordering on the
top level because apparently the files so far have all their keys
alphabetically ordered but the export from grafana is not. I doubt this
will be a long-term sustainable approach though.

Related progress issue: https://progress.opensuse.org/issues/122842