Avoid misleading result size graph for particular job groups
When selecting a particular job group in the "Approximate result
size by job group" panel to view only the result size for one
particular job group the Y-axis is changed to no longer start at
zero. I find graphs like this very confusing as they make small
changes appear much more dramatically as they are.

I came across this problem when investigating
https://progress.opensuse.org/issues/129244. During the
estimation we have added 'in particular check "Development
Security" which looks too big in too short a time' as suggestion
to that ticket. However, when changing the graph so the Y-axis
starts at zero this change doesn't look very big anymore which
is closer to the truth.