monitoring: Try to make silencing of "InfluxDB not reachable" effective
The alert for "InfluxDB not reachable" was still firing on reboots of
monitor.qa.suse.de even though we created a silence and configured a
special notification policy. Maybe we should treat the execution errors
as an actual "alert" with the setting "execErrState: Alerting" instead
of an "Error" which is apparently treated differently in grafana.

Related progress issue: https://progress.opensuse.org/issues/128870