Add host specific configuration to enable i2c rtc clock module
This enables a DS1307 hardware rtc module connected over i2c on our
piworker which is based on a raspberry pi 4. We refrain to do the
inital, one-time configuration here. So if this is enabled for another
host with a new rtc module one must sync the inital time with the help
of `hwclock -w` after an initial reboot.

Please also take a look at
https://progress.opensuse.org/issues/128942#note-26 for reference.