Fix calculated amount raid devices on retries
Currently we reset $raid_devices to 1 after each retry. If the previous
run however contained more than one device in the $device variable this
will cause mdadm to complain because it missmatches the amount of
supplied devices later on. Because we always want a RAID over *all*
devices, we can move the calculation right into the argument
construction to avoid having "out of sync" variables.