Allow boot if mount for /var/lib/openqa fails
Currently if the openqa_nvme_format.service fails for whatever reason,
the boot aborts and just spawns an early rescue shell. In reality we
would rather prefer if the system just continues to boot and connect to
our monitoring infrastructure so we can better resolve these issues.

I tested this manually on openqaworker1 and could confirm that it also
blocks openqa-worker instances to start.