Fix testing for telegraf errors
We grep for E! which means in case there are telegraf errors, the grep succeeds
and the pipeline will pass.
If there are no telegraf errors, grep will fail and the pipeline will fail.
So the logic is the opposite of what we want, and I can see actual telegraf
errors in older pipelines which passed.
Now that we fixed all telegraf errors, it fails.

This fix will check whether any errors landed in the logfile or not.