Make grub2-mkconfig runs in states idempotent
I realized that we execute grub2-mkconfig every time we run an
highstate. Despite the documentation stating different it seems like the
listen directive does not work like we understand it. With onchanges the
command behaves like expected and only executes if related files are
changed by salt. This also shaves of 2s execution time on every worker.