CI: Prevent missing logs on gitlab CI timeout
When gitlab CI jobs run into the hardcoded 2h timeout no logs are
uploaded and we do not really know where something was stuck or just
taking long. This commit adds a timeout call to the two longest-running
remote salt calls well below the gitlab CI 2h timeout so that we have at
least partial information preserved. If this is not enough then we might
need to potentially move the timeout call to the inner shell of ssh or
duplicate it there with lower timeout.

Related progress issue: https://progress.opensuse.org/issues/133457