Provide webUI jobs data in telegraf with optional webui_domain
Previously our FQDN for the main SUSE openQA instance was
"openqa.suse.de", same as we reference the domain in grafana monitoring
fed by telegraf. After moving the domain to a different FQDN and
changing openqa.suse.de to be a CNAME the monitoring data for "openQA
jobs" and alike are not showing up anymore under the same reference. To
fix that we can provide an optional grain "primary_webui_domain" and
fallback to FQDN as in before.

I already set the grain "primary_webui_domain" on OSD accordingly.

Related progress issue: https://progress.opensuse.org/issues/134816