Remove existing salt related locks before downgrading
If a lock for a previous version of a salt related package already
exists, zypper will still try to upgrade it (despite the <=3005 beeing
already fulfilled). This obviously fails because a lock is already in
place. We assume that existing locks without poo# reference are invalid
and therefore remove them before upgrading to the latest *known good*
version.