Align openvswitch config lookup with keys used in workers.ini
Our openvswitch state tries to automatically find all workers/instances
with the tap class configured. The keys used in the template of
workers.ini are based on the `host` grain in salt while openvswitch.sls
used the `nodename` grain to match on these keys. This was no problem
until now because `nodename`==`host` was true on all our workers. On
petrol the two grains differ and therefore cause a missing configuration
in our gre_tunnel_preup.sh script. This commit aligns this lookup to
rely on the same grain.