alerting: Reduce "Too few jobs" threshold according to recent observations
Originally the problem we observed was that "nearly no jobs are
executed" meaning 0 to about 3 jobs so the limit of 20 is a bit too
much. Depending on worker class matching we can not always ensure that
enough fitting worker ressources are available at a time for any worker
class if the scheduled jobs are only relying on those. Hence this commit
lowers the alert threshold from 20 to 5 as agreed in the daily QE Tools
infra call.

Related progress issue: https://progress.opensuse.org/issues/137603