Ensure proper SUSE CA repos on all salt controlled hosts
We previously only installed the SUSE CA repos on openQA worker or webui
role but not on other machines. That now lead to a problem on s390zl13
which was automatically redirecting downloads with "wget" to
https://openqa.suse.de which needs SUSE CA repos properly setup. Hence
this commit adds the according instructions to setup the repository and
package to install on all machines in the already existing
"ca-certificates" state.

Related progress issue: https://progress.opensuse.org/issues/152008