monitoring: Remove redundant and misleading graph for "scheduled+blocked"
The panel "Queue: State (SUSE)" is showing a stacked view of "running",
"scheduled", "blocked". The query "scheduled+blocked" was necessary for
creating a proper alert query which nowadays is separately managed with
the introduction of "unified alerting". The query caused a misleading
impression that 2x the actual number of jobs are scheduled or blocked
due to the redundancy of calculating the sum and then displaying the
result stacked on the individual graphs for scheduled and blocked.

This commit removes the redundant query so that the panel view provides
consistent results.