monitoring: Remove obsolete availability panel
The availability panel is an old copy from the home dashboard. Better
not keep old outdated copies around :)

This actually recovers the JSON export of the "status overview"
dashboard which in before I erroneously overwrote with the JSON model of
the home dashboard which AFAIK we otherwise don't save in the repo.