Avoid warning about `/etc/fstab` state
This state is merely there to ensure `/etc/fstab` exists. So it makes sense
that no source is specified. This leads to the warning:

```
State for file: /etc/fstab - Neither 'source' nor 'contents' nor
'contents_pillar' nor 'contents_grains' was defined, yet 'replace' was set
to 'True'. As there is no source to replace the file with, 'replace' has
been set to 'False' to avoid reading the file unnecessarily.
```

This change sets replace to false to avoid the warning. Note that
`file.exists` cannot be used because this function does not create the file
if it doesn't exist.

Related ticket: https://progress.opensuse.org/issues/158913