Fix location of worker_rlimit_nofile
as all files from conf.d are imported into the "http" section of the
nginx config the worker_rlimit_nofile setting was not accepted. Instead
we need to define worker_rlimit_nofile in the main config file which is
acceptable as we maintain the full file in salt anyway.

Verified with `nginx -t`

Related progress issue: https://progress.opensuse.org/issues/160239