nginx: Add X-Build-Tx-Time header to upstream
Currently OpenQA attaches this header itself once it receives a request.
If one of these requests is not processed fast enough, the backend might
reject authentication. This is an attempt to attach this header earlier
and more independent of the (maybe slow) backend application.

Also see: https://progress.opensuse.org/issues/162038