Introduce absolute memory usage alert for generic+workers
We found that depending on the workload a system runs the available
memory in either relative or absolute values matters for alerting us to
critical situations. This commit introduces an alert for the absolute
amount of available memory including swap below a threshold. I created
the alert definition first as a copy of the relative alert definition in
the grafana web interface, export that to a YAML file and then manually
incorporate the content into the existing alert definitions by copying
an existing entry and adjusting the entries as necessary.

Related progress issue: https://progress.opensuse.org/issues/164853