Drop sleperf config due to broken connectivity
This simply drops the sleperf states which currently can't work as the
connection to dashboard.qa2.suse.asia always times out to mitigate on-going
failures.

Note that the file is also renamed to avoid it being run as a unit test
despite not being or having a unit test.

See: https://progress.opensuse.org/issues/166235