Prevent unnecessary auto-update restarting to hide errors
In commit f6549bc I originally introduced Restart=on-failure however
that was for https://progress.opensuse.org/issues/62201 for problems to
access non-official repositories. For maintenance patches we should
actually assume a much higher level of availability of repositories so I
suggest to not retry there which might already fix the problem that the
service failure is not properly propagated.

Related progress issue: https://progress.opensuse.org/issues/166313